requirements.txt
streamlit
google-generativeai
pypdf
python-docx
SpeechRecognition
gTTS
streamlit-mic-recorder